authsystem/context.py
from datetime import timedelta
from datetime import datetime
from sqlalchemy.orm import Session

from jose import jwt
from passlib.context import CryptContext

from fastapi.security import OAuth2PasswordBearer

from authsystem.crud import (
    get_user_by_username as get_user_from_db_by_username,
)

from config import SECRET_KEY, ALGORITHM


pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

def verify_password(plain_password: str, user_password: str):
    return pwd_context.verify(plain_password, user_password)

def get_password_hash(password: str):
    return pwd_context.hash(password)

def get_user(db: Session, username: str):
    user = get_user_from_db_by_username(db=db, username=str(username))
    return user

def authenticate_user(db: Session, username: str, password: str):
    user = get_user(db, username)
    if not user:
        return False
    if not verify_password(password, user.password):
        return False
    return user

def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
